Set up events on the calendar

+++ app/src/main/java/com/example/sports_match_day/model/network/Match.kt
@@ -22,5 +22,5 @@ class Location(
 
 class Participant(
     var id: Int,
-    var score: Int
+    var score: Double
 )
